+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add functions
+    - `Serialize.toJson`
+    - `Serialize.toJsonPretty`
+    - `Serialize.dateTime`
+    - `Serialize.dateTimeOffset`
+    - `Serialize.stringOrNull`
 
-## 0.1.0 - 
+## 0.1.0 -
 - Initial implementation